# delilah_Resto_Project/project/api.yaml
swagger: "2.0"
info:
  description: "This is the Delilah Restó Restaurant API. Where you can make your order online, and receive it in your home, easy and quick."
  version: "1.0.0"
  title: "Delilah Restó"
  termsOfService: "http://delilah/terms/"
  contact:
    email: "delilahresto@gmail.com.co"
  license:
    name: "Apache 2.0"
    url: "http://www.delilahresto.com.co"
host: "delilah.swagger.io"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
...
  /users:
    get:
      tags:
      - "users"
      summary: "Get an specific user if not manager or get all user if is manager"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        type: "string"
        name: "token"
        description: "Token to authenticate"
        required: true
      responses:
       "200":
        description: "successful operation"
        schema:
          type: "array"
          items:
            $ref: "#/definitions/User"
       "403":
        description: "session expired"
        schema:
          type: "array"
          items:
            $ref: "#/definitions/Expired"
    post:
      tags:
      - "users"
      summary: "Create a new user (user sign up)"
      description: ""
      operationId: "createUsersWithArrayInput"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "JSON Object with all user information."
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/UserToPost"
      responses:
       "200":
        description: "Successful operation"
        schema:
          type: "array"
          items:
            $ref: "#/definitions/signUp"
       "400":
        description: "user_name, full_name, email, phone, address and password are required."
        schema:
          type: "array"
          items:
            $ref: "#/definitions/alreadyExist"    
       "409":
        description: "Username or email don't availables, User already exists."
        schema:
          type: "array"
          items:
            $ref: "#/definitions/alreadyExist"      
    put:
      tags:
      - "users"
      summary: "Update an user (phone, address, email)"
      description: ""
      operationId: "updateUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "JSON object with properties to update, one or all properties that want to update."
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/UserToUpdate"
      - in: "header"
        type: "string"
        name: "token"
        description: "Token to authenticate"
        required: true    
      responses:
       "200":
        description: "User has been updated"
        schema:
          type: "string"
       "401":
        description: "Insufficient privileges to update users. Please remove id property."
        schema:
          type: "string"
       "403":
        description: "Session expired or User do not have permissions to access."
        schema:
          type: "array"
          items:
            $ref: "#/definitions/Expired"
       "404":
          description: "User does not exist"
          schema:
           type: "string"      
...
  /users/login:
    post:
      tags:
      - "users"
      summary: "Registered user login"
      description: ""
      operationId: "createUsersWithListInput"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "JSON Object with password and username or email"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/login"
      responses:
        "200":
          description: "token"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/token"
        "401":
          description: "User does not exist."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Expired"      
  /users/changePassword:
    put:
      tags:
      - "users"
      summary: "Change user password"
      description: ""
      operationId: "loginUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "password"
        description: "The new password to update."
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/password"
      - in: "header"
        type: "string"
        name: "token"
        description: "Token to authenticate."
        required: true      
      responses:
        "200":
          description: "Password has been updated."
          schema:
            type: "string"
        "400":
          description: "Password length must be greater than 3."
          schema:
            type: "string"    
        "403":
          description: "Session expired or User do not have permissions to access."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Expired" 
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      user_name:
        type: "string"
      full_name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "integer"
        format: "int64"
      address:
        type: "string"
      password:
        type: "string"
      admin:
        type: "boolean"
        default: false
        description: "User Status"
    xml:
      name: "User"
  UserToPost:
    type: "object"
    properties:
      user_name:
        type: "string"
      full_name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "integer"
        format: "int64"
      address:
        type: "string"
      password:
        type: "string"
    xml:
      name: "UserToPost"
  UserToUpdate:
    type: "object"
    properties:
      id:
        type: "integer"
      phone:
        type: "string"
      address:
        type: "string"
      email:
        type: "string"
    xml:
      name: "UserToUpdate"    
  Expired:
    type: "object"
    properties:
      error:
        type: "string"
        format: "int64"
    xml:
      name: "Expired"
  alreadyExist:
    type: "object"
    properties:
      error:
       type: "string"
       format: "int64"
    xml:
      name: "alreadyExist"    
  signUp:
    type: "object"
    properties:
      user_id:
        type: "integer"
        format: "int64"
    xml:
      name: "signUp"
  login:
    type: "object"
    properties:
      user_name:
        type: "string"
      password:
        type: "string"
    xml:
      name: "login"
  token:
    type: "object"
    properties:
      token:
        type: "string"
    xml:
      name: "token"
  password:
    type: "object"
    properties:
      password:
        type: "string"
    xml:
      name: "pasword"
...